Moving all command handlers to appropriate file
