+++ srcs/methodHandlers/handlePong.cpp
@@ -1,0 +1,8 @@
+#include "../../includes/Server.hpp"
+
+void Server::handlePong(const Command &command)
+{
+//	TODO
+	if (command.getTextPart() == std::to_string(command.getUser().getTimestamp()))
+		logger.logUserMessage("nice ", command.getUser(), INFO );
+}
