+++ srcs/methodHandlers/handleQuit.cpp
@@ -1,0 +1,6 @@
+#include "../../includes/Server.hpp"
+
+void Server::handleQuit(const Command &command) {
+	User &user = command.getUser();
+	killUser(user);
+}
